+++ e2e/selectors/assets/DetectedTokensView.selectors.js
